move generateCommand() to the correct simple.go
