+++ process/simple.go
@@ -72,18 +72,6 @@ func NewCommand(commandLine []string, workingDirectory string, env []string) (*C
 	}, nil
 }
 
-func (task *TaskRun) generateCommand(index int) error {
-	var err error
-	task.Commands[index], err = process.NewCommand(task.Payload.Command[index], taskContext.TaskDir, task.EnvVars())
-	if err != nil {
-		return err
-	}
-	task.logMux.RLock()
-	defer task.logMux.RUnlock()
-	task.Commands[index].DirectOutput(task.logWriter)
-	return nil
-}
-
 func (c *Command) SetEnv(envVar, value string) {
 	c.Cmd.Env = append(c.Cmd.Env, envVar+"="+value)
 }
